requirements.txt
numpy
matplotlib
Pillow
librosa
ultralytics
torch
fastapi